Moved survey utils into separate directory, refactor
